# docs/openapi.json
{
 "components": {
  "schemas": {
...
   "CustomTopology": {
    "properties": {
     "chs": {
      "maximum": 5.0,
      "minimum": 1.0,
      "title": "Chs",
      "type": "integer"
     },
     "cms_per_ch": {
      "maximum": 8.0,
      "minimum": 1.0,
      "title": "Cms Per Ch",
      "type": "integer"
     }
    },
    "required": [
     "chs",
     "cms_per_ch"
    ],
    "title": "CustomTopology",
    "type": "object"
   },
   "DeviceDetail": {
    "properties": {
     "device": {
      "$ref": "#/components/schemas/DeviceOut"
     },
     "sessions": {
      "items": {
       "$ref": "#/components/schemas/SessionOut"
      },
      "title": "Sessions",
      "type": "array"
     },
     "status_history": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "title": "Status History",
      "type": "array"
     }
    },
    "required": [
     "device",
     "sessions",
     "status_history"
    ],
    "title": "DeviceDetail",
    "type": "object"
   },
   "DeviceOut": {
    "properties": {
     "cluster": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cluster"
     },
     "designated": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Designated"
     },
     "designation_state": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Designation State"
     },
     "epoch": {
      "title": "Epoch",
      "type": "integer"
     },
     "ident": {
      "title": "Ident",
      "type": "string"
     },
     "pu_fingerprint": {
      "title": "Pu Fingerprint",
      "type": "string"
     },
     "role": {
      "title": "Role",
      "type": "string"
     },
     "status": {
      "title": "Status",
      "type": "string"
     },
     "undelivered": {
      "default": 0,
      "title": "Undelivered",
      "type": "integer"
     }
    },
    "required": [
     "ident",
     "role",
     "cluster",
     "status",
     "epoch",
     "pu_fingerprint"
    ],
    "title": "DeviceOut",
    "type": "object"
   },
...
   "FrameOut": {
    "properties": {
     "bytes_len": {
      "title": "Bytes Len",
      "type": "integer"
     },
     "checks": {
      "items": {
       "$ref": "#/components/schemas/CheckOut"
      },
      "title": "Checks",
      "type": "array"
     },
     "dst": {
      "title": "Dst",
      "type": "string"
     },
     "fate": {
      "title": "Fate",
      "type": "string"
     },
     "frame_id": {
      "title": "Frame Id",
      "type": "integer"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "job_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Job Id"
     },
     "label": {
      "title": "Label",
      "type": "string"
     },
     "paper_bits": {
      "title": "Paper Bits",
      "type": "integer"
     },
     "payload_hex": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Payload Hex"
     },
     "reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reason"
     },
     "run_tag": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Run Tag"
     },
     "sent_step": {
      "title": "Sent Step",
      "type": "integer"
     },
     "src": {
      "title": "Src",
      "type": "string"
     },
     "step": {
      "title": "Step",
      "type": "integer"
     },
     "to": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "To"
     },
     "verdict": {
      "title": "Verdict",
      "type": "string"
     }
    },
    "required": [
     "id",
     "frame_id",
     "job_id",
     "step",
     "sent_step",
     "src",
     "dst",
     "to",
     "label",
     "bytes_len",
     "paper_bits",
     "payload_hex",
     "verdict",
     "reason",
     "fate",
     "checks"
    ],
    "title": "FrameOut",
    "type": "object"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
   "JobAccepted": {
    "properties": {
     "created": {
      "title": "Created",
      "type": "boolean"
     },
     "job_id": {
      "title": "Job Id",
      "type": "integer"
     }
    },
    "required": [
     "job_id",
     "created"
    ],
    "title": "JobAccepted",
    "type": "object"
   },
   "JobCreate": {
    "properties": {
     "args": {
      "additionalProperties": true,
      "default": {},
      "title": "Args",
      "type": "object"
     },
     "idempotency_key": {
      "maxLength": 64,
      "minLength": 8,
      "pattern": "^[A-Za-z0-9_-]+$",
      "title": "Idempotency Key",
      "type": "string"
     },
     "type": {
      "enum": [
       "onboard",
       "step",
       "send_readings",
       "start_periodic",
       "stop_periodic",
       "rekey",
       "revoke",
       "reprovision",
       "designate",
       "lab_scenario",
       "evaluation",
       "reset_network"
      ],
      "title": "Type",
      "type": "string"
     }
    },
    "required": [
     "type",
     "idempotency_key"
    ],
    "title": "JobCreate",
    "type": "object"
   },
...
   "NetworkCreate": {
    "properties": {
     "custom": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/CustomTopology"
       },
       {
        "type": "null"
       }
      ]
     },
     "forward_ciphertexts": {
      "default": false,
      "title": "Forward Ciphertexts",
      "type": "boolean"
     },
     "kind": {
      "default": "product",
      "enum": [
       "product",
       "lab"
      ],
      "title": "Kind",
      "type": "string"
     },
     "mode": {
      "default": "enhanced",
      "enum": [
       "enhanced",
       "original"
      ],
      "title": "Mode",
      "type": "string"
     },
     "name": {
      "pattern": "^[A-Za-z0-9][A-Za-z0-9 _.-]{0,63}$",
      "title": "Name",
      "type": "string"
     },
     "params": {
      "default": "demo",
      "enum": [
       "toy",
       "demo",
       "secure"
      ],
      "title": "Params",
      "type": "string"
     },
     "secure_pseudo_ids": {
      "default": true,
      "title": "Secure Pseudo Ids",
      "type": "boolean"
     },
     "seed": {
      "anyOf": [
       {
        "maximum": 4.611686018427388e+18,
        "minimum": 0.0,
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Seed"
     },
     "template": {
      "anyOf": [
       {
        "enum": [
         "paper",
         "small",
         "net"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": "paper",
      "title": "Template"
     }
    },
    "required": [
     "name"
    ],
    "title": "NetworkCreate",
    "type": "object"
   },
   "NetworkDelete": {
    "properties": {
     "confirm_name": {
      "title": "Confirm Name",
      "type": "string"
     }
    },
    "required": [
     "confirm_name"
    ],
    "title": "NetworkDelete",
    "type": "object"
   },
   "NetworkDetail": {
    "properties": {
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "device_count": {
      "title": "Device Count",
      "type": "integer"
     },
     "devices": {
      "items": {
       "$ref": "#/components/schemas/DeviceOut"
      },
      "title": "Devices",
      "type": "array"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "kind": {
      "title": "Kind",
      "type": "string"
     },
     "mode": {
      "title": "Mode",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "options": {
      "additionalProperties": true,
      "default": {},
      "title": "Options",
      "type": "object"
     },
     "params": {
      "title": "Params",
      "type": "string"
     },
     "security_level": {
      "title": "Security Level",
      "type": "string"
     },
     "seed": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Seed"
     },
     "status": {
      "title": "Status",
      "type": "string"
     },
     "status_counts": {
      "additionalProperties": {
       "type": "integer"
      },
      "title": "Status Counts",
      "type": "object"
     },
     "step": {
      "title": "Step",
      "type": "integer"
     },
     "template": {
      "title": "Template",
      "type": "string"
     },
     "tunables": {
      "additionalProperties": {
       "type": "integer"
      },
      "title": "Tunables",
      "type": "object"
     }
    },
    "required": [
     "id",
     "name",
     "kind",
     "mode",
     "params",
     "security_level",
     "template",
     "seed",
     "status",
     "step",
     "created_at",
     "status_counts",
     "device_count",
     "devices",
     "tunables"
    ],
    "title": "NetworkDetail",
    "type": "object"
   },
   "NetworkOut": {
    "properties": {
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "device_count": {
      "title": "Device Count",
      "type": "integer"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "kind": {
      "title": "Kind",
      "type": "string"
     },
     "mode": {
      "title": "Mode",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "options": {
      "additionalProperties": true,
      "default": {},
      "title": "Options",
      "type": "object"
     },
     "params": {
      "title": "Params",
      "type": "string"
     },
     "security_level": {
      "title": "Security Level",
      "type": "string"
     },
     "seed": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Seed"
     },
     "status": {
      "title": "Status",
      "type": "string"
     },
     "status_counts": {
      "additionalProperties": {
       "type": "integer"
      },
      "title": "Status Counts",
      "type": "object"
     },
     "step": {
      "title": "Step",
      "type": "integer"
     },
     "template": {
      "title": "Template",
      "type": "string"
     }
    },
    "required": [
     "id",
     "name",
     "kind",
     "mode",
     "params",
     "security_level",
     "template",
     "seed",
     "status",
     "step",
     "created_at",
     "status_counts",
     "device_count"
    ],
    "title": "NetworkOut",
    "type": "object"
   },
...
   "Page_FrameOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/FrameOut"
      },
      "title": "Items",
      "type": "array"
     },
     "next_cursor": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Cursor"
     }
    },
    "required": [
     "items"
    ],
    "title": "Page[FrameOut]",
    "type": "object"
   },
   "Page_ReadingOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ReadingOut"
      },
      "title": "Items",
      "type": "array"
     },
     "next_cursor": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Cursor"
     }
    },
    "required": [
     "items"
    ],
    "title": "Page[ReadingOut]",
    "type": "object"
   },
   "ReadingOut": {
    "properties": {
     "device": {
      "title": "Device",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "received_step": {
      "title": "Received Step",
      "type": "integer"
     },
     "seq": {
      "title": "Seq",
      "type": "integer"
     },
     "value": {
      "title": "Value"
     }
    },
    "required": [
     "id",
     "device",
     "seq",
     "value",
     "received_step"
    ],
    "title": "ReadingOut",
    "type": "object"
   },
...
   "SessionOut": {
    "properties": {
     "a": {
      "title": "A",
      "type": "string"
     },
     "b": {
      "title": "B",
      "type": "string"
     },
     "epoch": {
      "title": "Epoch",
      "type": "integer"
     },
     "established_step": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Established Step"
     },
     "purpose": {
      "title": "Purpose",
      "type": "string"
     },
     "recv": {
      "title": "Recv",
      "type": "integer"
     },
     "sent": {
      "title": "Sent",
      "type": "integer"
     },
     "sid_hex": {
      "title": "Sid Hex",
      "type": "string"
     },
     "state": {
      "title": "State",
      "type": "string"
     },
     "superseded_by": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Superseded By"
     }
    },
    "required": [
     "a",
     "b",
     "purpose",
     "state",
     "sid_hex",
     "established_step",
     "epoch",
     "sent",
     "recv"
    ],
    "title": "SessionOut",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "MAKA Secure IoT Console",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/networks": {
   "get": {
    "operationId": "list_networks_api_networks_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/NetworkOut"
         },
         "title": "Response List Networks Api Networks Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "List Networks",
    "tags": [
     "networks"
    ]
   },
   "post": {
    "operationId": "create_network_api_networks_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/NetworkCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NetworkDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Network",
    "tags": [
     "networks"
    ]
   }
  },
  "/api/networks/{network_id}": {
   "delete": {
    "operationId": "delete_network_api_networks__network_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "network_id",
      "required": true,
      "schema": {
       "title": "Network Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/NetworkDelete"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Network",
    "tags": [
     "networks"
    ]
   },
   "get": {
    "operationId": "get_network_api_networks__network_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "network_id",
      "required": true,
      "schema": {
       "title": "Network Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NetworkDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Network",
    "tags": [
     "networks"
    ]
   }
  },
  "/api/networks/{network_id}/devices/{ident}": {
   "get": {
    "operationId": "get_device_api_networks__network_id__devices__ident__get",
    "parameters": [
     {
      "in": "path",
      "name": "network_id",
      "required": true,
      "schema": {
       "title": "Network Id",
       "type": "integer"
      }
     },
     {
      "in": "path",
      "name": "ident",
      "required": true,
      "schema": {
       "title": "Ident",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DeviceDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Device",
    "tags": [
     "networks"
    ]
   }
  },
  "/api/networks/{network_id}/frames": {
   "get": {
    "operationId": "list_frames_api_networks__network_id__frames_get",
    "parameters": [
     {
      "in": "path",
      "name": "network_id",
      "required": true,
      "schema": {
       "title": "Network Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "label",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Label"
      }
     },
     {
      "in": "query",
      "name": "src",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Src"
      }
     },
     {
      "in": "query",
      "name": "dst",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Dst"
      }
     },
     {
      "in": "query",
      "name": "verdict",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Verdict"
      }
     },
     {
      "in": "query",
      "name": "after_step",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "After Step"
      }
     },
     {
      "in": "query",
      "name": "job",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Job"
      }
     },
     {
      "in": "query",
      "name": "run_tag",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Run Tag"
      }
     },
     {
      "in": "query",
      "name": "cursor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 200,
       "maximum": 1000,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_FrameOut_"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Frames",
    "tags": [
     "networks"
    ]
   }
  },
  "/api/networks/{network_id}/jobs": {
   "post": {
    "operationId": "submit_job_api_networks__network_id__jobs_post",
    "parameters": [
     {
      "in": "path",
      "name": "network_id",
      "required": true,
      "schema": {
       "title": "Network Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/JobCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobAccepted"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Submit Job",
    "tags": [
     "networks"
    ]
   }
  },
  "/api/networks/{network_id}/readings": {
   "get": {
    "description": "FR-07: decrypted readings at the BS -- Operator and Admin only (V-WEB-08).",
    "operationId": "list_readings_api_networks__network_id__readings_get",
    "parameters": [
     {
      "in": "path",
      "name": "network_id",
      "required": true,
      "schema": {
       "title": "Network Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "device",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Device"
      }
     },
     {
      "in": "query",
      "name": "cursor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 200,
       "maximum": 1000,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_ReadingOut_"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Readings",
    "tags": [
     "networks"
    ]
   }
  },
  "/api/networks/{network_id}/stream": {
   "get": {
    "description": "SSE: frame, security_event, device_status, job_progress. Resumes after Last-Event-ID.\n`once=true` returns the buffered events and closes (used by tests and polling clients).",
    "operationId": "stream_api_networks__network_id__stream_get",
    "parameters": [
     {
      "in": "path",
      "name": "network_id",
      "required": true,
      "schema": {
       "title": "Network Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "last_event_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Last Event Id"
      }
     },
     {
      "in": "query",
      "name": "once",
      "required": false,
      "schema": {
       "default": false,
       "title": "Once",
       "type": "boolean"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Stream",
    "tags": [
     "networks"
    ]
   }
  }
 }
}